refactor(bloater): remove SearchBar from Header and create a separated component
